# vcnc-rest/api/swagger.yaml
...
swagger: '2.0'
...
info:
  version: "1.1.0"
  title: Velstor vCNC API
  description: |
    ### Admin API

    The Admin API manages the vCNC server itself.

    ### vTRQ API

    The vTRQ API performs operations on a vTRQ.

    Errors are reported at two levels: HTTP and operational.  With one exception, HTTP error codes represent the success or failure of the HTTP protocol, considered solely as a transport. Operational error codes, reported in the response body, reflect the success or failure of the operation itself.

    The exception is the operational error code __ENOENT__ (Does not exist).  REST operations which refer to exactly one resource in the URL convert __ENOENT__ to HTTP 404.  Doing so improves our compliance with [*HTTP/1.1*](http://tools.ietf.org/html/rfc7231).  Refer to the [vCNC API Overview](../api-overview.html) for a more detailed discussion.

# Generic host; our customers deploy this service.
# In production, comment out this line.  The correct host will be displayed
# when the deployed vCNC server hosts this document set.
#host: api.acme.com

# Array of all schemes the API supports.
schemes:
  - http

# will be prefixed to all paths
basePath: /v1
...
consumes:
  - application/json
produces:
  - application/json

#
#  Endpoint definitions
#
paths:
...
  /grid/jobs/{job_id}:
    post:
      summary: Creates a new resource describing a job with the specified id.
      operationId: gridJobPost
      tags:
        - "Grid API"
      parameters:
        - in: path
          name: job_id
          description: ID of grid job
          required: true
          type: string
        - in: body
          name: body
          description: Information about a grid job
          required: true
          schema:
            type: object
            properties:
              workspace_name:
                type: string
              workspace_spec:
                type: string
            required:
              - workspace_name
              - workspace_spec
            additionalProperties: false
      responses:
        '200':
          description: Request processed.  See response body.
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'

    get:
      summary: Retrieves data associated with a grid job.
      operationId: gridJobGet
      tags:
        - "Grid API"
      produces:
        - application/json
      parameters:
        - in: path
          name: job_id
          description: ID of grid job
          required: true
          type: string
      responses:
        '200':
          description: Request processed.  See response body.
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'

    delete:
      summary: Deletes data associated with a grid job.
      operationId: gridJobDelete
      tags:
        - "Grid API"
      parameters:
        - in: path
          name: job_id
          description: ID of grid job
          required: true
          type: string
      responses:
        '200':
          description: Request processed.  See response body.
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  ##
  ##  vCNC - vtrq
  ##
    post:
      summary: Creates a new vTRQ resource and assigns the vTRQ ID.
      operationId: vcncVtrqCreate
      tags:
        - "Admin API"
      parameters:
        - in: body
          name: body
          description: "Information about a host machine."
          required: true
          schema:
            $ref: '#/definitions/Host'
        - in: header
          name: "Authorization"
          description: JWT session token
          required: true
          type: string
      responses:
        '200':
          description: Request processed.  See response body.
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'

    put:
      summary: Updates a vTRQ entry.
      operationId: vcncVtrqUpdate
      tags:
        - "Admin API"
      parameters:
        - in: path
          name: vtrq_id
          description: ID of vTRQ
          required: true
          type: integer
          format: int64
          minimum: 0
          maximum: 1073741823
        - in: body
          name: body
          description: "Information about a host machine."
          required: true
          schema:
            $ref: '#/definitions/Host'
        - in: header
          name: "Authorization"
          description: JWT session token
          required: true
          type: string
      responses:
        '200':
          description: Request processed.  See response body.
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'

    get:
      summary: Retrieves data associated with a vTRQ.
      operationId: vcncVtrqGet
      tags:
        - "Admin API"
      produces:
        - application/json
      parameters:
        - in: path
          name: vtrq_id
          description: ID of vTRQ
          required: true
          type: integer
          format: int64
          minimum: 0
          maximum: 1073741823
        - in: header
          name: "Authorization"
          description: JWT session token
          required: true
          type: string
      responses:
        '200':
          description: Request processed.  See response body.
          schema:
            $ref: '#/definitions/Host'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'

    delete:
      summary: Deletes a vTRQ entry.
      operationId: vcncVtrqDelete
      tags:
        - "Admin API"
      parameters:
        - in: path
          name: vtrq_id
          description: ID of vTRQ
          required: true
          type: integer
          format: int64
          minimum: 0
          maximum: 1073741823
        - in: header
          name: "Authorization"
          description: JWT session token
          required: true
          type: string
      responses:
        '200':
          description: Request processed.  See response body.
        '400':
          description: Ill-formed or invalid request.
          schema:
            $ref: '#/definitions/Error'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'

  ##
  ##  vCNC - users
  ##
...
  /vtrq/delete_nodes/{vtrq_id}:
    post:
      summary: Performs a vector of deletions in a vTRQ namespace
      tags:
        - "vTRQ Namespace"
      operationId: vtrqDeleteNodes
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: vtrq_id
          description: ID of the vTRQ performing this operation.
          required: true
          type: integer
          format: int64
          minimum: 0
          maximum: 1073741823
        - in: query
          name: recursive
          required: false
          type: boolean
          default: false
        - in: body
          name: body
          description: "Array of deletion requests."
          required: true
          schema:
            $ref: '#/definitions/DeleteNodesRequest'
      responses:
        '200':
          description: Request processed, see response body.
          schema:
            $ref: '#/definitions/DeleteNodesResponse'
        default:
          description: HTTP error
          schema:
            $ref: '#/definitions/Error'
  /vtrq/meta_copy/{vtrq_id}:
    post:
      summary: Performs a vector of meta-data copies
      description: |
        Performs internal "meta-data" copies of one part of the vTRQ namespace to another.

        If the source path is a directory and the destination path exists, the copy is not performed and the error code is set to __EEXIST__.

        If the source path is a file and the destination path is an existing directory, the copy is not performed and the error code is set to __EEXIST__.

        When both the source path and destination path are existing files, the behavior depends on the value of "overwrite". If true, the file at the destination path is atomically changed to be the file at the source path. Otherwise, the copy is not performed and the error code is set to __EEXIST__.

        Metadata copies are performed on the server in no particular order. The return JSON object pairs each metacopy request with its resulting error code. The array is in no particular order; specifically, it may or may not be in the order it was executed on the vTRQ.

      tags:
        - "vTRQ Namespace"
      operationId: vtrqMetaCopy
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: vtrq_id
          description: ID of the vTRQ performing this operation.
          required: true
          type: integer
          format: int64
          minimum: 0
          maximum: 1073741823
        - in: query
          name: overwrite
          required: false
          type: boolean
          default: false
        - in: body
          name: body
          description: "Array of meta-copy requests."
          required: true
          schema:
            $ref: '#/definitions/MetaCopyRequest'
      responses:
        '200':
          description: Request processed, see response body.
          schema:
            $ref: '#/definitions/MetaCopyResponse'
        default:
          description: HTTP error
          schema:
            $ref: '#/definitions/Error'
  /vtrq/namespace/{vtrq_id}/{url_path}:
    delete:
      summary: Deletes the node at the specified paath.
      operationId: vtrqNamespaceDelete
      tags:
        - "vTRQ Namespace"
      parameters:
        - in: path
          name: vtrq_id
          description: ID of the vTRQ performing this operation.
          required: true
          type: integer
          format: int64
          minimum: 0
          maximum: 1073741823
        - in: path
          name: url_path
          description: URL-encoded path in the vTRQ namespace.
          required: true
          type: string
          format: url-path
        - in: query
          name: recursive
          required: false
          type: boolean
          default: false
      responses:
        '200':
          description: Request processed.  See response body.
          schema:
            $ref: '#/definitions/ErrorVtrq'
        default:
          description: HTTP error
          schema:
            $ref: '#/definitions/Error'

    get:
      summary: Returns meta-data about the specified path
      description: "The returned content resembles that of 'struct stat' on Unix systems."

      operationId: vtrqNamespaceGet
      tags:
        - "vTRQ Namespace"
      parameters:
        - in: path
          name: vtrq_id
          description: ID of the vTRQ performing this operation.
          required: true
          type: integer
          format: int64
          minimum: 0
          maximum: 1073741823
        - in: path
          name: url_path
          description: URL-encoded path in the vTRQ namespace.
          required: true
          type: string
          format: url-path
      responses:
        '200':
          description: Request processed.  See response body.
          schema:
            required:
              - error_sym
            properties:
              error_sym:
                type: string
                description: Symbolic error code value
                default: OK
                enum:
                  - OK
                  - EACCESS
                  - EREMOTEIO
              stat:
                $ref: '#/definitions/StructStat'
        default:
          description: HTTP error
          schema:
            $ref: '#/definitions/Error'
  /vtrq/namespace/{vtrq_id}/{url_path}/children:
    get:
      summary: Returns the child nodes of a namespace path
      operationId: vtrqNamespaceChildren
      tags:
        - "vTRQ Namespace"
      parameters:
        - in: path
          name: vtrq_id
          description: ID of the vTRQ performing this operation.
          required: true
          type: integer
          format: int64
          minimum: 0
          maximum: 1073741823
        - in: path
          name: url_path
          description: URL-encoded path in the vTRQ namespace.
          required: true
          type: string
        - in: query
          name: filter
          required: false
          type: string
        - in: query
          name: limit
          description: Number of items to retreive. Default is 1000; maximum is 100,000.
          required: false
          type: integer
          default: 1000
        - in: query
          name: cursor
          description: Opaque value for start of next set of items. Zero means start at beginning.
          required: false
          type: integer
          default: 0
        - in: query
          name: details
          description: If false the 'stat' field will not be returned.
          required: false
          type: boolean
          default: false
      responses:
        '200':
          description: Request processed.  See response body.
          schema:
            type: object
            required:
              - error_sym
              - result
            properties:
              error_sym:
                type: string
                default: OK
              result:
                type: array
                items:
                  type: string
            additionalProperties: false
        default:
          description: HTTP error
          schema:
            $ref: '#/definitions/Error'
  /vtrq/namespace/{vtrq_id}/{url_path}/consistency:
    get:
      summary: Retreives the consistency attribute of a vTRQ node.
      operationId: vtrqNamespaceConsistencyGet
      tags:
        - "vTRQ Namespace"
      parameters:
        - in: path
          name: vtrq_id
          description: ID of the vTRQ performing this operation.
          required: true
          type: integer
          format: int64
          minimum: 0
          maximum: 1073741823
        - in: path
          name: url_path
          description: URL-encoded path in the vTRQ namespace.
          required: true
          type: string
      responses:
        '200':
          description: Request processed. See response body.
          schema:
            required:
              - error_sym
              - consistency
            properties:
              error_sym:
                type: string
                description: Symbolic error code value
                default: OK
                enum:
                  - OK
                  - EACCESS
                  - ENOENT
                  - ENOTDIR
                  - ENAMETOOLONG
                  - EEXIST
                  - ENOSPC
                  - EREMOTEIO
              consistency:
                $ref: '#/definitions/Consistency'
            additionalProperties: false
        default:
          description: HTTP error
          schema:
            $ref: '#/definitions/Error'

    post:
      summary: Sets the consistency attribute of a vTRQ node.
      operationId: vtrqNamespaceConsistencySet
      tags:
        - "vTRQ Namespace"
      parameters:
        - in: path
          name: vtrq_id
          description: ID of the vTRQ performing this operation.
          required: true
          type: integer
          format: int64
          minimum: 0
          maximum: 1073741823
        - in: path
          name: url_path
          description: URL-encoded path in the vTRQ namespace.
          required: true
          type: string
        - in: body
          name: body
          description: "Additional parameters"
          required: true
          schema:
            type: object
            properties:
              consistency:
                $ref: '#/definitions/Consistency'
            required:
              - consistency
            additionalProperties: false
      responses:
        '200':
          description: Request processed. See response body.
          schema:
            required:
              - error_sym
            properties:
              error_sym:
                type: string
                description: Symbolic error code value
                default: OK
                enum:
                  - OK
                  - EACCESS
                  - ENOENT
                  - ENOTDIR
                  - ENAMETOOLONG
                  - EEXIST
                  - ENOSPC
                  - EREMOTEIO
            additionalProperties: false
        default:
          description: HTTP error
          schema:
            $ref: '#/definitions/Error'
  /vtrq/namespace/{vtrq_id}/{url_path}/mkdir:
    post:
      summary: Creates a directory at a namespace path.
      description: Clients must ensure the parent exists and is a directory.
      operationId: vtrqNamespaceMkdir
      tags:
        - "vTRQ Namespace"
      parameters:
        - in: path
          name: vtrq_id
          description: ID of the vTRQ performing this operation.
          required: true
          type: integer
          format: int64
          minimum: 0
          maximum: 1073741823
        - in: path
          name: url_path
          description: URL-encoded path in the vTRQ namespace.
          required: true
          type: string
        - in: query
          name: mode
          description: "File access bits (decimal)"
          required: false
          type: integer
          format: int32
          default: 429
        - in: query
          name: parents
          description: When 'true', creates parent directories as necessary.
          required: false
          type: boolean
          default: false
      responses:
        '200':
          description: Request processed. See response body.
          schema:
            required:
              - error_sym
            properties:
              error_sym:
                type: string
                description: Symbolic error code value
                default: OK
                enum:
                  - OK
                  - EACCESS
                  - ENOENT
                  - ENOTDIR
                  - ENAMETOOLONG
                  - EEXIST
                  - ENOSPC
                  - EREMOTEIO
        default:
          description: HTTP error
          schema:
            $ref: '#/definitions/Error'
  /vtrq/service/{vtrq_id}:
    delete:
      summary: Gracefully brings down a vTRQ
      operationId: vtrqShutdown
      tags:
        - "vTRQ Services"
      parameters:
        - in: path
          name: vtrq_id
          description: ID of the vTRQ performing this operation.
          required: true
          type: integer
          format: int64
          minimum: 0
          maximum: 1073741823
      responses:
        '202':
          description: Request accepted. Client may receive the HTTP response before the vTRQ has fully shut down.
          schema:
            properties:
              error_sym:
                type: string
                description: Symbolic error code value
                default: OK
                enum:
                  - OK
            required:
              - error_sym
            additionalProperties: false
        default:
          description: HTTP error
          schema:
            $ref: '#/definitions/Error'
...
  /vtrq/vp/{vtrq_id}:
    get:
      summary: Finds VP instances matching certain criteria
      description: |
        The result body contains the id of every VP which matches all (logical AND) of the search parameters. Matching is based on simple string equality.

        If the request body contains no parameters, every VP id is returned.

      operationId: vpFind
      tags:
        - "VP Instances"
      parameters:
        - in: path
          name: vtrq_id
          description: ID of the vTRQ performing this operation.
          required: true
          type: integer
        - in: query
          name: vp_host
          description: Hostname of the machine running the VP
          required: false
          type: string
        - in: query
          name: mount_point
          description: Fully qualified canonical path to the VP mount point.
          required: false
          type: string
        - in: query
          name: limit
          description: Number of items to retreive. Default is 1000; maximum is 100,000.
          required: false
          type: integer
          default: 1000
        - in: query
          name: cursor
          description: Opaque value for start of next set of items. Zero means start at beginning.
          required: false
          type: integer
          default: 0
        - in: query
          name: details
          required: false
          type: boolean
          default: false
      responses:
        '200':
          description: Request processed, see response body.
          schema:
            required:
              - error_sym
            properties:
              error_sym:
                type: string
                description: Symbolic error code value
                default: OK
                enum:
                  - OK
                  - EACCESS
                  - EREMOTEIO
              vp_ids:
                type: array
                description: ""
                items:
                  type: string
                  example: "0x123456789ABCDEF0"
        default:
          description: HTTP error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  DeleteNodesRequest:
    type: object
    properties:
      delete_paths:
        type: array
        items:
          type: string
          example: /path/on/vtrq_host
    required:
      - delete_paths
    additionalProperties: false
...
  Error:
    required:
      - error_sym
      - message
    properties:
      error_sym:
        type: string
      message:
        type: string
...
  Host:
    type: object
    properties:
      hname:
        type: string
      vtrq_id:
        type: integer
      port:
        type: string
      transport:
        type: string
        default: tcp
        enum:
          - tcp
          - sctp
    required:
      - hname
      - vtrq_id
      - port
      - transport
    additionalProperties: false
...
  SourceDestinationVector:
    type: array
    items:
      $ref: '#/definitions/SourceDestinationPath'
...
  MetaCopyRequest:
    type: object
    properties:
      copy_paths:
        $ref: '#/definitions/SourceDestinationVector'
    required:
      - copy_paths
    additionalProperties: false
...